backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
email-validator==2.1.0
python-dateutil==2.8.2

# File handling
pillow==10.2.0
aiofiles==23.2.1

# Testing (optional for later)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

